+++ app/src/main/java/com/example/opscpoe3/Category.kt
@@ -1,0 +1,7 @@
+package com.example.opscpoe3
+
+
+data class Category(
+    val name: String,
+    val description: String,
+    val color: Int)
